+++ packages/awino-ui/src/components/pages/landing/StatsSection/StatsSection.tsx
@@ -1,15 +1,10 @@
-import { useCallback, useState } from 'react';
-
-import CountUp from 'react-countup';
-import VisibilitySensor from 'react-visibility-sensor';
-
-import { Container, Grid, Typography } from '@mui/material';
+import { Typography } from '@mui/material';
 import { styled } from '@mui/material/styles';
-import trimEnd from 'lodash/trimEnd';
 
 import Section from '@/components/layout/Section/Section';
 import usePageTranslation from '@/hooks/usePageTranslation';
 import { StatsData } from '@/types/pages/landing';
+import StatsItems from '../../shared/StatsItems/StatsItems';
 
 const Root = styled(Section)(({ theme }) => ({
   textAlign: 'center',
@@ -35,17 +30,6 @@ const Root = styled(Section)(({ theme }) => ({
       backgroundColor: theme.palette.background.transparent,
     },
   },
-  '.card': {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    maxWidth: '300px',
-    height: '100%',
-    padding: theme.spacing(12, 7, 11),
-    margin: '0 auto',
-    borderRadius: +theme.shape.borderRadius * 5,
-    backgroundColor: theme.palette.background.transparent,
-  },
   [theme.breakpoints.up('md')]: {
     h1: {
       fontSize: '5rem' /* 80px */,
@@ -61,16 +45,6 @@ interface Props {
 }
 export default function StatsSection({ items }: Props) {
   const t = usePageTranslation();
-  const [glanced, setGlanced] = useState(false);
-
-  const handleVisibilityChange = useCallback(
-    (isVisible: boolean) => {
-      if (!glanced && isVisible) {
-        setGlanced(true);
-      }
-    },
-    [glanced, setGlanced]
-  );
 
   return (
     <>
@@ -83,61 +57,7 @@ export default function StatsSection({ items }: Props) {
           ))}
         </ul>
         <Typography variant="h1">{t('stats-section.title')}</Typography>
-        <VisibilitySensor
-          active={!glanced}
-          partialVisibility
-          offset={{ bottom: 100 }}
-          onChange={handleVisibilityChange}
-        >
-          <Container maxWidth="lg">
-            <Grid container spacing={8}>
-              {items.map(({ value, subvalue }, index) => {
-                const formatter = (value: number) => {
-                  return t(`stats-section.items.${index}.value`, { v: trimEnd(`${value}`, '0') });
-                };
-                return (
-                  <Grid key={index} item xs={12} sm={6} lg={3}>
-                    <div className="card">
-                      <h2>
-                        <Typography
-                          variant="h3"
-                          fontWeight={600}
-                          component="span"
-                          sx={{
-                            display: 'block',
-                            mb: 3,
-                          }}
-                        >
-                          {glanced ? (
-                            <CountUp delay={0} end={value} duration={2} decimals={2} formattingFn={formatter}>
-                              {({ countUpRef }) => <span ref={countUpRef} />}
-                            </CountUp>
-                          ) : (
-                            <span>&nbsp;</span>
-                          )}
-                        </Typography>
-                        <Typography
-                          variant="body-md"
-                          fontWeight={500}
-                          component="span"
-                          color="text.primary"
-                          sx={{ display: 'block', mb: 1.5 }}
-                        >
-                          {t(`stats-section.items.${index}.title`)}
-                        </Typography>
-                      </h2>
-                      {subvalue && (
-                        <Typography variant="body-sm" fontWeight={500}>
-                          {t(`stats-section.items.${index}.subvalue`, { v: subvalue })}
-                        </Typography>
-                      )}
-                    </div>
-                  </Grid>
-                );
-              })}
-            </Grid>
-          </Container>
-        </VisibilitySensor>
+        <StatsItems items={items} maxWidth="lg" />
       </Root>
     </>
   );