+++ packages/awino-ui/src/pages/market.tsx
@@ -6,20 +6,23 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { setPageI18nNamespace } from '@/app/state/slices/app';
 import storeWrapper from '@/app/store';
 import Seo from '@/components/layout/Seo/Seo';
-import StatsSection from '@/components/pages/landing/StatsSection/StatsSection';
+import StatsSection from '@/components/pages/market/StatsSection/StatsSection';
 
 const stats = [
   { value: 89.7, subvalue: 24.72 },
   { value: 89.7, subvalue: 24.72 },
   { value: 0.27 },
   { value: 273.4, subvalue: 52 },
+  { value: 47.4, subvalue: 24.72 },
 ];
 
-const IndexPage: NextPage = () => {
+const assets = [];
+const MarketPage: NextPage = () => {
   return (
     <>
       <Seo />
       <StatsSection items={stats} />
+      {/* <AssetSection items={assets} /> */}
     </>
   );
 };
@@ -35,4 +38,4 @@ export const getServerSideProps = storeWrapper.getServerSideProps((store) => asy
   };
 });
 
-export default IndexPage;
+export default MarketPage;